Zakończenie bazowej części treningowej
Pełen podstawowy etap mechanizmu dla treningu zrobiony, dla czesci pełnej i skróconej. Dodano parę nowych funkcji i usprawnień widoku.

+++ lib/database_classes/DatabaseHelper.dart
@@ -48,8 +48,8 @@ class DatabaseHelper {
           id INTEGER PRIMARY KEY AUTOINCREMENT,
           exercises TEXT,
           date TEXT,
-          intensity INTEGER,
-          satisfaction INTEGER,
+          intensity REAL,
+          satisfaction REAL,
           comment TEXT,
           time INTEGER
         )
